Implemented parsing of almost all richtext tags recognized by Unity UI and TextMeshPro.
Richtext values are placed in TextConfiguration. Many of them in a stack (as done by TextMeshPro), which enables enable nesting. It is up to the layout engine to use these metrics correctly when generating the glyphs, lines, and pages.  Adopting the layout engine of TextMeshPro would be several thousand lines of code and is not done in this pull request.