Version 0.12.0 Alpha.13 [Prerelease]
Warning: This commit hash may not be preserved!

Compared to the previous version, this version fixes various issues with editor world state interacting with live baking which would often lead to incorrect previews or other errors. In addition, it fixes a baking error with optimized skeletons in newer Unity versions, and it fixes an issue with masked animation sampling being flipped.